+++ Plugin/Catalog/Block/ListProduct.php
@@ -39,9 +39,8 @@ namespace Nosto\Cmp\Plugin\Catalog\Block;
 use Magento\Catalog\Block\Product\ListProduct as MagentoListProduct;
 use Magento\Catalog\Model\Product;
 use Magento\Catalog\Model\ResourceModel\Product\Collection;
-use Nosto\Cmp\Model\Service\Recommendation\StateAwareCategoryServiceInterface;
+use Nosto\Cmp\Model\Service\Merchandise\LastResult;
 use Nosto\Cmp\Plugin\Catalog\Model\Product as NostoProductPlugin;
-use Nosto\Cmp\Utils\CategoryMerchandising;
 
 /**
  * This interceptor is used to pass `nosto_cmp` parameter to product urls when they are sorted by Nosto
@@ -50,14 +49,14 @@ use Nosto\Cmp\Utils\CategoryMerchandising;
 class ListProduct
 {
     /**
-     * @var StateAwareCategoryServiceInterface
+     * @var LastResult
      */
-    private $categoryService;
+    private $lastResult;
 
     public function __construct(
-        StateAwareCategoryServiceInterface $categoryService
+        LastResult $lastResult
     ) {
-        $this->categoryService = $categoryService;
+        $this->lastResult = $lastResult;
     }
 
     /**
@@ -70,10 +69,10 @@ class ListProduct
         MagentoListProduct $listProduct,
         Collection $collection
     ) {
-        $categoryMerchandisingResult = $this->categoryService->getLastResult();
+        $categoryMerchandisingResult = $this->lastResult->getLastResult();
 
         if ($categoryMerchandisingResult != null) {
-            $cmpProductIds = CategoryMerchandising::parseProductIds($categoryMerchandisingResult);
+            $cmpProductIds = $categoryMerchandisingResult->parseProductIds();
             $collection->each(static function ($product) use ($cmpProductIds) {
                 /* @var Product $product */
                 if (in_array($product->getId(), $cmpProductIds)) {